requirements.txt
Django==3.1.4
django-bootstrap4==2.3.1
Pillow==8.2.0
psycopg2-binary==2.8.6
netifaces==0.10.9
djangorestframework==3.12.2
sentry-sdk==0.19.5
beautifulsoup4==4.9.3
requests==2.26.0